+++ packages/sheets-data-validation/src/plugin.ts
@@ -38,7 +38,6 @@ import { defaultPluginConfig, PLUGIN_CONFIG_KEY } from './controllers/config.sch
 import { DataValidationController } from './controllers/dv.controller';
 import { DataValidationFormulaController } from './controllers/dv-formula.controller';
 import { DataValidationFormulaRefRangeController } from './controllers/dv-formula-ref-range.controller';
-import { DataValidationRefRangeController } from './controllers/dv-ref-range.controller';
 import { SheetDataValidationSheetController } from './controllers/dv-sheet.controller';
 import { SheetDataValidationModel } from './models/sheet-data-validation-model';
 import { DataValidationCacheService } from './services/dv-cache.service';
@@ -72,7 +71,6 @@ export class UniverSheetsDataValidationPlugin extends Plugin {
             [SheetsDataValidationValidatorService],
             [SheetDataValidationModel],
             [DataValidationController],
-            [DataValidationRefRangeController],
             [DataValidationFormulaController],
             [SheetDataValidationSheetController],
             [DataValidationFormulaRefRangeController],
@@ -94,7 +92,6 @@ export class UniverSheetsDataValidationPlugin extends Plugin {
 
         this._injector.get(DataValidationCacheService);
         this._injector.get(SheetsDataValidationValidatorService);
-        this._injector.get(DataValidationRefRangeController);
         this._injector.get(DataValidationFormulaRefRangeController);
     }
